restructured config.yaml(s) fixed path problems in analysis_pipeline caused by the restructure
same with ui and sever files

sample deletion now work with a minor bug if the user deletes the only sample that they have there is a shiny error, investigating the reasons.

boxes are no longer collapsible there were issues with css

basic error check with deletion and progress bar, it usally too fast for the progress to be visible but might change in production.